ParkWhere/backend/requirements.txt
django 
requests 
djangorestframework
simple-history